Everything works

+++ manga-downloader/main.py
@@ -20,19 +20,22 @@ def scan_chapter(link):
     browser.get(link)
     select = Select(browser.find_element_by_id('chapterSelectorSelect'))
     option = select.first_selected_option
-    print('Creating folder: ' + option.text)
-    os.mkdir(option.text)
-    os.chdir(option.text)
-    print('Folder created')
-    count = browser.find_element_by_class_name('pages-count').text
-    for counter in range(int(count)):
-        print('Downloading ' + str(counter + 1) + ' from ' + count + ' in ' + option.text)
-        imgUrl = browser.find_element_by_id('mangaPicture').get_attribute('src')
-        urllib.request.urlretrieve(imgUrl, str(counter))
-        browser.find_element_by_class_name('fa-arrow-right').click()
-        time.sleep(5)
-    browser.quit()
-    os.chdir('..')
+    if os.path.exists(option.text):
+        print("Skip chapter: " + option.text)
+    else:
+        print('Creating folder: ' + option.text)
+        os.mkdir(option.text)
+        os.chdir(option.text)
+        print('Folder created')
+        count = browser.find_element_by_class_name('pages-count').text
+        for counter in range(int(count)):
+            print('Downloading ' + str(counter + 1) + ' from ' + count + ' in ' + option.text)
+            imgUrl = browser.find_element_by_id('mangaPicture').get_attribute('src')
+            urllib.request.urlretrieve(imgUrl, str(counter))
+            browser.find_element_by_class_name('fa-arrow-right').click()
+            time.sleep(5)
+        browser.quit()
+        os.chdir('..')
 
 
 def clean_name(text):
@@ -69,7 +72,8 @@ if __name__ == '__main__':
     rate = int(input("Enter the rate of manga (0 or 1 (where 1 is a 18+)): "))
     get_chapter_links(mainLink)
     if os.path.exists('chaptesLinks.txt'):
+        num_lines = sum(1 for line in open('chaptesLinks.txt'))
         with open('chaptesLinks.txt') as file:
-            lines = file.readlines();
+            lines = file.readlines()
         for line in lines:
             scan_chapter(line)
